+++ apx_core/src/url/mod.rs
@@ -1,5 +1,6 @@
 //! URL types
 
+pub mod ap_uri;
 pub mod canonical;
 pub mod common;
 pub mod hostname;